Rename DepositApi file
