+++ src/api/index.ts
@@ -6,7 +6,7 @@ import { TokenListApiMock } from './tokenList/TokenListApiMock'
 import { Erc20ApiMock } from './erc20/Erc20ApiMock'
 import { Erc20ApiImpl, Erc20Api } from './erc20/Erc20Api'
 import { DepositApiMock } from './exchange/DepositApiMock'
-import { DepositApiImpl, DepositApi } from './exchange/DepositApiImpl'
+import { DepositApiImpl, DepositApi } from './exchange/DepositApi'
 import { ExchangeApiImpl, ExchangeApi } from './exchange/ExchangeApiImpl'
 import { ExchangeApiMock } from './exchange/ExchangeApiMock'
 import {
